python/requirements.txt
black==23.10.1
isort==5.12.0
mypy==1.6.1
pip==23.3.1
pyproject-autoflake==1.0.2
pyproject-flake8==6.1.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
setuptools==68.2.2